Renamed Model Program
Renamed generate_model to more accurately reflect what it does.
